Refactor core logic to a separate crate
It will allow other crate to use this as a library
