+++ Cargo.toml
@@ -26,9 +26,8 @@ path = "src/generate_tests.rs"
 [workspace]
 
 [dependencies]
-proc-macro2 = "1.0"
-quote = "1.0"
 schemafy_core = { version = "0.5.0", path = "schemafy_core" } # VERSION_TAG
+schemafy_lib = { version = "0.5.0", path = "schemafy_lib" }   # VERSION_TAG
 serde = "1.0"
 serde_json = "1.0"
 serde_derive = "1.0"
