+++ schemafy_lib/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "schemafy_lib"
+version = "0.5.0" # VERSION_TAG
+authors = ["Markus Westerlind <marwes91@gmail.com>", "MATILLAT Quentin <qmatillat@gmail.com>"]
+edition = "2018"
+
+[dependencies]
+proc-macro2 = "1.0"
+quote = "1.0"
+schemafy_core = { version = "0.5.0", path = "../schemafy_core" } # VERSION_TAG
+serde = "1.0"
+serde_json = "1.0"
+serde_derive = "1.0"
+syn = "1.0"
+
+Inflector = "0.11"
+
